Added new bottom navigation bar and Created separate class files for home page, scheduler page and settings page
