+++ lib/pages/settings.dart
@@ -1,0 +1,52 @@
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:firebase_database/firebase_database.dart';
+
+
+class settingsPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: settingspage(),
+      debugShowCheckedModeBanner: false,
+    );
+  }
+}
+
+class settingspage extends StatefulWidget {
+  const settingspage({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() {
+    return _settingspage();
+  }
+}
+
+class _settingspage extends State<settingspage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          "Settings",
+          style: TextStyle(
+              color: Colors.black, fontWeight: FontWeight.bold, fontSize: 24),
+        ),
+        backgroundColor: Colors.blue,
+        centerTitle: true,
+      ),
+      body: Container(
+        alignment: Alignment.topCenter,
+        padding: const EdgeInsets.all(20),
+        decoration: const BoxDecoration(
+          image: DecorationImage(
+              image: AssetImage("images/background.jpg"), fit: BoxFit.cover),
+        ),
+        child: Column(
+          children: [],
+        ),
+      ),
+    );
+  }
+}
